bugfix read log file in admin log screen

+++ app/controllers/admin/log_controller.rb
@@ -1,14 +1,14 @@
 class Admin::LogController < Admin::Base
-def index
-  @fname = 'log/task.log'
-  @loglines = Array.new  
-  file = File.open(@fname)
-  if (file.present?) then
-    while text = file.gets do
-      text.chomp!
-      @loglines.push(text)
+  def index
+    @fname = Rails.root.join('log', 'task.log')
+    @loglines = Array.new
+    if File.exist?(@fname) then
+      file = File.open(@fname)
+      while text = file.gets do
+        text.chomp!
+        @loglines.push(text)
+      end
+      file.close
     end
-    file.close
   end
-end
 end
